+++ src/components/Header/NavbarMenu/UserMenu.tsx
@@ -10,7 +10,7 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
+} from "@/components/ui/DropdownMenu"
 
 type UserMenuProps = {
   session: Session
